+++ components/Form/Picker.jsx
@@ -4,6 +4,7 @@ import { ModalForm } from './Modal'
 export const PickerSelect = ({
   placeholder, grow, value, onChange, children, title, showButton, disabled, childPropsValueKey,
   _designKey,
+  modalFormProps,
   ...props
 }) => {
 
@@ -17,6 +18,7 @@ export const PickerSelect = ({
     onChange={onChange}
     disabled={disabled}
     childPropsValueKey={childPropsValueKey}
+    {...modalFormProps}
     _designKey={_designKey}
   />
 }
@@ -24,6 +26,7 @@ export const PickerSelect = ({
 export const PickerMultiSelect = ({
   placeholder, grow, value, onChange, children, title, showButton, disabled, childPropsValueKey,
   _designKey,
+  modalFormProps,
   ...props
 }) => {
 
@@ -37,6 +40,7 @@ export const PickerMultiSelect = ({
     onChange={onChange}
     disabled={disabled}
     childPropsValueKey={childPropsValueKey}
+    {...modalFormProps}
     _designKey={_designKey}
   />
 }
@@ -49,6 +53,7 @@ PickerMultiSelect.defaultProps = {
 export const PickerDate = ({
   placeholder, grow, value, onChange, children, title, showButton, disabled, childPropsValueKey,
   _designKey,
+  modalFormProps,
   ...props
 }) => {
   return <ModalForm
@@ -62,5 +67,6 @@ export const PickerDate = ({
     disabled={disabled}
     childPropsValueKey={childPropsValueKey}
     _designKey={_designKey}
+    {...modalFormProps}
   >{children}</ModalForm>
 }